FIX: extracting out o1js from snarkjs
